Add sharing functionality and code refactoring
Sharing is done by simply appending the list "sharedWith" on a profile on FireStore.

Used switch statement when receiving messages in background.js.

Moved toast function to background.js as both buttons use it. Also moved most of the styles to css. Also also included status bar for the toast